Изменил тип данных поля mass

+++ src/main/java/com/new_ton/domain/dto/accountmanager/AccountManagerTableDataDto.java
@@ -13,13 +13,13 @@ public class AccountManagerTableDataDto {
     private String brend;
     private String nameprod;
     private Double percent;
-    private Integer mass;
+    private Double mass;
     private Double tempprodmin;
     private Double tempprodmax;
     private String comment;
     private String strDate;
 
-    public AccountManagerTableDataDto(Integer idpr, Date datecr, String brend, String nameprod, Double percent, Integer mass, Double tempprodmin, Double tempprodmax, String comment) {
+    public AccountManagerTableDataDto(Integer idpr, Date datecr, String brend, String nameprod, Double percent, Double mass, Double tempprodmin, Double tempprodmax, String comment) {
         this.idpr = idpr;
         this.datecr = datecr;
         this.brend = brend;
